% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_limno.R
\name{scale_color_limno}
\alias{scale_color_limno}
\title{scale_color_limno}
\usage{
scale_color_limno(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
scale_color_limno
}
